+++ takin-web-data/src/main/java/io/shulie/takin/web/data/dao/tracenode/TraceNodeDao.java
@@ -34,7 +34,7 @@ public interface TraceNodeDao {
      *
      * @return
      */
-    Integer getNodeCount(String traceId, Long customerId);
+    Long getNodeCount(String traceId, Long customerId);
 
     /**
      * 异常节点个数
